test: Refactor for pytest's `--import-mode=importlib`
According to the pytest documentation, this setting will become the
default in the future.
