--- tests/__init__.py
@@ -1,8 +1,0 @@
-# SPDX-FileCopyrightText: Copyright DB Netz AG and the capellambse contributors
-# SPDX-License-Identifier: Apache-2.0
-
-import pathlib
-import re
-
-TEST_ROOT = pathlib.Path(__file__).parent / "data" / "melodymodel"
-TEST_MODEL = "Melody Model Test.aird"
