+++ tests/test_model_layers.py
@@ -25,7 +25,7 @@ from capellambse.model.layers.ctx import SystemComponentPkg
 from capellambse.model.layers.la import CapabilityRealization
 from capellambse.model.layers.oa import OperationalCapability
 
-from . import TEST_MODEL, TEST_ROOT
+from .conftest import TEST_MODEL, TEST_ROOT
 
 
 def test_model_info_contains_capella_version(model: MelodyModel):
